cnn_with_policy_gradients/csci5525_project/requirements.txt
numpy
torch
matplotlib
gym==0.21.0
gym-retro
importlib-metadata==4.13.0
baselines
imageio